@@ -1,54 +1,24 @@
-import './App.css';
-import './Nav.css';
-import Header from'./components/Header';
-import MainBody from './components/MainBody';
-import AdminsApp from './AdminApp/AdminApp';
-import VendorsApp from './VendorsApp/VendorsApp';
-import CustomersApp from './CustomersApp/CustomersApp';
-import MenuSelection from './CustomersApp/MenuSelection';
+import "./App.css"
+import "./Nav.css"
+import AdminsApp from "./AdminApp/AdminApp"
+import VendorsApp from "./VendorsApp/VendorsApp"
+import CustomersApp from "./CustomersApp/CustomersApp"
+import MenuSelection from "./CustomersApp/MenuSelection"
 
-import React, { Component } from 'react';
-import {
-    BrowserRouter as Router,
-    Switch,
-    Route,
-    Link
-  } from "react-router-dom";
-
+import React, { Component } from "react"
+import { Routes, Route } from "react-router-dom"
 
 function App() {
-
-  
-
   return (
-  
-    <div>  
-      <Router>
-      {/* A <Switch> looks through its children <Route>s and
-          renders the first one that matches the current URL. */}
-        <Switch>
-        <Route path="/menu">
-            {/* menu selection component */}
-            <MenuSelection />
-          </Route>
-          <Route path="/admin">
-            {/* adminpage component */}
-            <AdminsApp />
-          </Route>
-          <Route path="/vendor">
-            {/* vendorpage component */}
-            <VendorsApp />
-            </Route>
-          <Route path="/">
-            {/* customerpage component */}
-            <CustomersApp />
-          </Route>
-        </Switch>
-      </Router>
+    <div>
+      <Routes>
+        <Route path="/menu" element={<MenuSelection />} />
+        <Route path="/admin/*" element={<AdminsApp />} />
+        <Route path="/vendor/*" element={<VendorsApp />} />
+        <Route path="/" element={<CustomersApp />} />
+      </Routes>
     </div>
-
   )
-
 }
 
-export default App;
+export default App
